Address tech debt: remove dead exports, split execution-review, extract dashboard helpers
- Remove 12 dead exports across 5 lib files (deterministic.ts, day-items.ts,
  types.ts, config.ts) that had zero imports outside their defining file
- Split lib/execution-review.ts (1,295 lines) into three modules:
  - execution-review-types.ts: type definitions and Zod schemas
  - execution-review-persistence.ts: Supabase persistence and weekly brief
  - execution-review.ts: core evidence-building and AI verdict generation (713 lines)
- Extract 15 helper functions and 10 types from dashboard/page.tsx (1,449 lines)
  into dashboard-helpers.ts, reducing page.tsx to 1,048 lines

All existing imports continue to work via re-exports from the original module paths.
