+++ lib/execution-review.ts
@@ -1,337 +1,33 @@
 import "openai/shims/node";
-import type { SupabaseClient } from "@supabase/supabase-js";
 import { zodTextFormat } from "openai/helpers/zod";
-import { z } from "zod";
 import { asObject, asString, asStringArray, clip } from "@/lib/openai";
 import { callOpenAIWithFallback } from "@/lib/ai/call-with-fallback";
 import type { AthleteContextSnapshot } from "@/lib/athlete-context";
-import { diagnoseCompletedSession, type SessionDiagnosisInput, type Sport } from "@/lib/coach/session-diagnosis";
+import { diagnoseCompletedSession, type SessionDiagnosisInput } from "@/lib/coach/session-diagnosis";
+import {
+  type ExecutionEvidence,
+  type CoachVerdict,
+  coachVerdictSchema,
+  COACH_VERDICT_JSON_EXAMPLE,
+} from "@/lib/execution-review-types";
+import {
+  nextCallFromEvidence,
+  buildEvidenceSummary,
+} from "@/lib/execution-review-persistence";
 
-export type ExecutionEvidence = {
-  sessionId: string;
-  athleteId: string;
-  sport: Sport;
-  planned: {
-    title: string;
-    intentCategory: string | null;
-    durationSec: number | null;
-    targetBands: {
-      hr?: { min?: number; max?: number };
-      power?: { min?: number; max?: number };
-      pace?: { min?: number; max?: number };
-    } | null;
-    plannedIntervals: number | null;
-    plannedStructure: string | null;
-    sessionRole: "key" | "supporting" | "recovery" | "unknown";
-  };
-  actual: {
-    durationSec: number | null;
-    avgHr: number | null;
-    avgPower: number | null;
-    /** Duration-weighted average power from work-interval laps only. */
-    avgIntervalPower?: number | null;
-    avgPaceSPerKm: number | null;
-    timeAboveTargetPct: number | null;
-    intervalCompletionPct: number | null;
-    variabilityIndex: number | null;
-    normalizedPower?: number | null;
-    trainingStressScore?: number | null;
-    intensityFactor?: number | null;
-    totalWorkKj?: number | null;
-    avgCadence?: number | null;
-    avgPacePer100mSec?: number | null;
-    bestPacePer100mSec?: number | null;
-    avgStrokeRateSpm?: number | null;
-    maxStrokeRateSpm?: number | null;
-    avgSwolf?: number | null;
-    elevationGainM?: number | null;
-    elevationLossM?: number | null;
-    poolLengthM?: number | null;
-    lengthCount?: number | null;
-    hrZoneTimeSec?: number | null;
-    paceZoneTimeSec?: number | null;
-    maxHr?: number | null;
-    maxPower?: number | null;
-    maxCadence?: number | null;
-    bestPaceSPerKm?: number | null;
-    normalizedGradedPaceSPerKm?: number | null;
-    aerobicTrainingEffect?: number | null;
-    anaerobicTrainingEffect?: number | null;
-    splitMetrics: {
-      firstHalfAvgHr?: number;
-      lastHalfAvgHr?: number;
-      firstHalfAvgPower?: number;
-      lastHalfAvgPower?: number;
-      firstHalfPaceSPerKm?: number;
-      lastHalfPaceSPerKm?: number;
-      firstHalfAvgCadence?: number;
-      lastHalfAvgCadence?: number;
-      firstHalfPacePer100mSec?: number;
-      lastHalfPacePer100mSec?: number;
-      firstHalfStrokeRate?: number;
-      lastHalfStrokeRate?: number;
-    } | null;
-    sportSpecific?: {
-      run?: {
-        avgPaceSPerKm: number | null;
-        bestPaceSPerKm: number | null;
-        normalizedGradedPaceSPerKm: number | null;
-        avgHr: number | null;
-        maxHr: number | null;
-        hrZoneTimeSec: number | null;
-        paceZoneTimeSec: number | null;
-        avgCadence: number | null;
-        maxCadence: number | null;
-        elevationGainM: number | null;
-        elevationLossM: number | null;
-        trainingStressScore: number | null;
-        aerobicTrainingEffect: number | null;
-        anaerobicTrainingEffect: number | null;
-        splitMetrics: ExecutionEvidence["actual"]["splitMetrics"];
-      } | null;
-      swim?: {
-        avgPacePer100mSec: number | null;
-        bestPacePer100mSec: number | null;
-        avgStrokeRateSpm: number | null;
-        maxStrokeRateSpm: number | null;
-        avgSwolf: number | null;
-        poolLengthM: number | null;
-        lengthCount: number | null;
-        paceZoneTimeSec: number | null;
-        trainingStressScore: number | null;
-        aerobicTrainingEffect: number | null;
-        anaerobicTrainingEffect: number | null;
-        splitMetrics: ExecutionEvidence["actual"]["splitMetrics"];
-      } | null;
-      bike?: {
-        avgPower: number | null;
-        normalizedPower: number | null;
-        maxPower: number | null;
-        intensityFactor: number | null;
-        variabilityIndex: number | null;
-        totalWorkKj: number | null;
-        avgCadence: number | null;
-        maxCadence: number | null;
-        avgHr: number | null;
-        maxHr: number | null;
-        hrZoneTimeSec: number | null;
-        trainingStressScore: number | null;
-        aerobicTrainingEffect: number | null;
-        anaerobicTrainingEffect: number | null;
-        splitMetrics: ExecutionEvidence["actual"]["splitMetrics"];
-      } | null;
-      strength?: {
-        durationSec: number | null;
-        intervalCompletionPct: number | null;
-        avgHr: number | null;
-        maxHr: number | null;
-        timeAboveTargetPct: number | null;
-        trainingStressScore: number | null;
-        aerobicTrainingEffect: number | null;
-        anaerobicTrainingEffect: number | null;
-      } | null;
-    } | null;
-  };
-  detectedIssues: Array<{
-    code: string;
-    severity: "low" | "moderate" | "high";
-    supportingMetrics: string[];
-  }>;
-  missingEvidence: string[];
-  rulesSummary: {
-    intentMatch: "on_target" | "partial" | "missed";
-    executionScore: number | null;
-    executionScoreBand: "On target" | "Solid" | "Partial match" | "Missed intent" | null;
-    confidence: "high" | "medium" | "low";
-    provisional: boolean;
-    evidenceCount: number;
-    executionCost: "low" | "moderate" | "high" | "unknown";
-  };
-};
-
-export type CoachVerdict = {
-  sessionVerdict: {
-    headline: string;
-    summary: string;
-    intentMatch: "on_target" | "partial" | "missed";
-    executionCost: "low" | "moderate" | "high" | "unknown";
-    confidence: "high" | "medium" | "low";
-    nextCall: "move_on" | "proceed_with_caution" | "repeat_session" | "protect_recovery" | "adjust_next_key_session";
-  };
-  explanation: {
-    sessionIntent?: string | null;
-    whatHappened: string;
-    whyItMatters: string;
-    oneThingToChange?: string | null;
-    whatToDoNextTime: string;
-    whatToDoThisWeek: string;
-  };
-  uncertainty: {
-    label: "confident_read" | "early_read" | "insufficient_data";
-    detail: string;
-    missingEvidence: string[];
-  };
-  citedEvidence: Array<{
-    claim: string;
-    support: string[];
-  }>;
-};
-
-export type WeeklyExecutionBrief = {
-  weekHeadline: string;
-  weekSummary: string;
-  keyPositive: string | null;
-  keyRisk: string | null;
-  nextWeekDecision: string;
-  trend: {
-    reviewedCount: number;
-    onTargetCount: number;
-    partialCount: number;
-    missedCount: number;
-    provisionalCount: number;
-  };
-  sessionsNeedingAttention: Array<{
-    sessionId: string;
-    sessionName: string;
-    scoreHeadline: string;
-    reason: string;
-  }>;
-  confidenceNote: string | null;
-};
-
-export type PersistedExecutionReview = {
-  version: 2;
-  linkedActivityId: string | null;
-  deterministic: ExecutionEvidence;
-  verdict: CoachVerdict | null;
-  narrativeSource: "ai" | "fallback" | "legacy_unknown";
-  weeklyImpact: {
-    suggestedWeekAction: string;
-    suggestedNextCall: "move_on" | "proceed_with_caution" | "repeat_session" | "protect_recovery" | "adjust_next_key_session";
-  } | null;
-  createdAt: string;
-  updatedAt: string;
-  status: "matched_intent" | "partial_intent" | "missed_intent";
-  intentMatchStatus: "matched_intent" | "partial_intent" | "missed_intent";
-  executionScore: number | null;
-  executionScoreBand: "On target" | "Solid" | "Partial match" | "Missed intent" | null;
-  executionScoreSummary: string;
-  executionSummary: string;
-  summary: string;
-  whyItMatters: string;
-  recommendedNextAction: string;
-  diagnosisConfidence: "high" | "medium" | "low";
-  executionScoreProvisional: boolean;
-  suggestedWeekAdjustment: string;
-  evidence: string[];
-  durationCompletion: number | null;
-  intervalCompletionPct: number | null;
-  timeAboveTargetPct: number | null;
-  avgHr: number | null;
-  avgPower: number | null;
-  avgIntervalPower?: number | null;
-  normalizedPower: number | null;
-  trainingStressScore: number | null;
-  intensityFactor: number | null;
-  totalWorkKj: number | null;
-  avgCadence: number | null;
-  avgPacePer100mSec?: number | null;
-  bestPacePer100mSec?: number | null;
-  avgStrokeRateSpm?: number | null;
-  maxStrokeRateSpm?: number | null;
-  avgSwolf?: number | null;
-  elevationGainM?: number | null;
-  elevationLossM?: number | null;
-  poolLengthM?: number | null;
-  lengthCount?: number | null;
-  hrZoneTimeSec?: number | null;
-  paceZoneTimeSec?: number | null;
-  maxHr: number | null;
-  maxPower: number | null;
-  firstHalfAvgHr: number | null;
-  lastHalfAvgHr: number | null;
-  firstHalfPaceSPerKm: number | null;
-  lastHalfPaceSPerKm: number | null;
-  firstHalfAvgCadence?: number | null;
-  lastHalfAvgCadence?: number | null;
-  firstHalfPacePer100mSec?: number | null;
-  lastHalfPacePer100mSec?: number | null;
-  firstHalfStrokeRate?: number | null;
-  lastHalfStrokeRate?: number | null;
-  executionCost: "low" | "moderate" | "high" | "unknown";
-  missingEvidence: string[];
-};
-
-const coachVerdictSchema = z.object({
-  sessionVerdict: z.object({
-    headline: z.string().min(1).max(160),
-    summary: z.string().min(1).max(500),
-    intentMatch: z.enum(["on_target", "partial", "missed"]),
-    executionCost: z.enum(["low", "moderate", "high", "unknown"]),
-    confidence: z.enum(["high", "medium", "low"]),
-    nextCall: z.enum(["move_on", "proceed_with_caution", "repeat_session", "protect_recovery", "adjust_next_key_session"])
-  }),
-  explanation: z.object({
-    sessionIntent: z.string().min(1).max(300).nullable().optional(),
-    whatHappened: z.string().min(1).max(500),
-    whyItMatters: z.string().min(1).max(500),
-    oneThingToChange: z.string().min(1).max(500).nullable().optional(),
-    whatToDoNextTime: z.string().min(1).max(500),
-    whatToDoThisWeek: z.string().min(1).max(500)
-  }),
-  uncertainty: z.object({
-    label: z.enum(["confident_read", "early_read", "insufficient_data"]),
-    detail: z.string().min(1).max(500),
-    missingEvidence: z.array(z.string().min(1)).max(8)
-  }),
-  citedEvidence: z.array(z.object({
-    claim: z.string().min(1).max(200),
-    support: z.array(z.string().min(1).max(180)).max(4)
-  })).max(4)
-});
-
-const COACH_VERDICT_EXAMPLE: CoachVerdict = {
-  sessionVerdict: {
-    headline: "Intent partially landed",
-    summary: "The session hit some of the intended stimulus, but late fade and incomplete work kept it from fully landing.",
-    intentMatch: "partial",
-    executionCost: "moderate",
-    confidence: "medium",
-    nextCall: "proceed_with_caution"
-  },
-  explanation: {
-    sessionIntent: "Threshold intervals to push lactate clearance at race-relevant intensity.",
-    whatHappened: "Intensity stayed 5% below target through the first 4 reps, then faded. The last 2 reps were incomplete.",
-    whyItMatters: "That means the session delivered some useful stimulus, but not the precise version the week was counting on.",
-    oneThingToChange: "NEXT threshold intervals: hold 165-175 bpm for all 6 reps. Success: complete the full set without fade. Start 2% easier on rep 1.",
-    whatToDoNextTime: "Start the first work block a touch easier so you can hold form and finish the full set.",
-    whatToDoThisWeek: "Keep the next key session on the calendar, but avoid forcing progression if fatigue is still lingering."
-  },
-  uncertainty: {
-    label: "early_read",
-    detail: "This read is useful, but it is still limited by missing split or interval detail.",
-    missingEvidence: ["split comparison"]
-  },
-  citedEvidence: [
-    {
-      claim: "Late fade showed up in the second half.",
-      support: ["Second-half pace slowed versus the first half", "Interval completion was below plan"]
-    }
-  ]
-};
-
-const COACH_VERDICT_JSON_EXAMPLE = JSON.stringify(COACH_VERDICT_EXAMPLE, null, 2);
+// Re-export types and persistence so existing imports from "@/lib/execution-review" keep working.
+export type { ExecutionEvidence, CoachVerdict, WeeklyExecutionBrief, PersistedExecutionReview } from "@/lib/execution-review-types";
+export {
+  toPersistedExecutionReview,
+  parsePersistedExecutionReview,
+  refreshObservedPatterns,
+  buildWeeklyExecutionBrief,
+} from "@/lib/execution-review-persistence";
 
 function toIntentMatch(status: "matched_intent" | "partial_intent" | "missed_intent") {
   if (status === "matched_intent") return "on_target" as const;
   if (status === "missed_intent") return "missed" as const;
   return "partial" as const;
-}
-
-function toLegacyStatus(status: "on_target" | "partial" | "missed") {
-  if (status === "on_target") return "matched_intent" as const;
-  if (status === "missed") return "missed_intent" as const;
-  return "partial_intent" as const;
 }
 
 function deriveMissingEvidence(input: SessionDiagnosisInput) {
@@ -780,44 +476,6 @@ function deriveExecutionCost(params: {
   return "moderate" as const;
 }
 
-function buildEvidenceSummary(evidence: ExecutionEvidence) {
-  const points: string[] = [];
-  if (evidence.actual.intervalCompletionPct !== null) {
-    points.push(`${Math.round(evidence.actual.intervalCompletionPct * 100)}% of planned reps completed`);
-  }
-  if (evidence.actual.timeAboveTargetPct !== null) {
-    points.push(`${Math.round(evidence.actual.timeAboveTargetPct * 100)}% above target`);
-  }
-  if (evidence.actual.avgHr !== null) {
-    points.push(`average HR ${Math.round(evidence.actual.avgHr)} bpm`);
-  }
-  if (typeof evidence.actual.avgIntervalPower === "number") {
-    points.push(`average interval power ${Math.round(evidence.actual.avgIntervalPower)} w`);
-  }
-  if (evidence.actual.avgPower !== null) {
-    const label = evidence.actual.avgIntervalPower != null ? "average power (session)" : "average power";
-    points.push(`${label} ${Math.round(evidence.actual.avgPower)} w`);
-  }
-  if (typeof evidence.actual.normalizedPower === "number") {
-    points.push(`normalized power ${Math.round(evidence.actual.normalizedPower)} w`);
-  }
-  if (typeof evidence.actual.variabilityIndex === "number") {
-    points.push(`variability index ${evidence.actual.variabilityIndex.toFixed(2)}`);
-  }
-  if (typeof evidence.actual.trainingStressScore === "number") {
-    points.push(`TSS ${Math.round(evidence.actual.trainingStressScore)}`);
-  }
-  return points.slice(0, 4);
-}
-
-function nextCallFromEvidence(intentMatch: "on_target" | "partial" | "missed", executionCost: "low" | "moderate" | "high" | "unknown") {
-  if (intentMatch === "on_target" && executionCost === "low") return "move_on" as const;
-  if (intentMatch === "missed" && executionCost === "high") return "protect_recovery" as const;
-  if (intentMatch === "missed") return "repeat_session" as const;
-  if (executionCost === "high") return "adjust_next_key_session" as const;
-  return "proceed_with_caution" as const;
-}
-
 function buildDeterministicVerdict(evidence: ExecutionEvidence): CoachVerdict {
   const intentMatch = evidence.rulesSummary.intentMatch;
   const nextCall = nextCallFromEvidence(intentMatch, evidence.rulesSummary.executionCost);
@@ -1053,243 +711,3 @@ export async function generateCoachVerdict(args: {
   return { verdict: result.value, source: result.source };
 }
 
-export function toPersistedExecutionReview(args: {
-  linkedActivityId: string | null;
-  evidence: ExecutionEvidence;
-  verdict: CoachVerdict | null;
-  narrativeSource?: "ai" | "fallback" | "legacy_unknown";
-  createdAt?: string;
-  updatedAt?: string;
-}): PersistedExecutionReview {
-  const createdAt = args.createdAt ?? new Date().toISOString();
-  const updatedAt = args.updatedAt ?? createdAt;
-  const legacyStatus = toLegacyStatus(args.evidence.rulesSummary.intentMatch);
-  const nextCall = args.verdict?.sessionVerdict.nextCall ?? nextCallFromEvidence(args.evidence.rulesSummary.intentMatch, args.evidence.rulesSummary.executionCost);
-  const suggestedWeekAction = args.verdict?.explanation.whatToDoThisWeek ?? "Keep the rest of the week stable and use this review as guidance for the next similar session.";
-
-  return {
-    version: 2,
-    linkedActivityId: args.linkedActivityId,
-    deterministic: args.evidence,
-    verdict: args.verdict,
-    narrativeSource: args.narrativeSource ?? (args.verdict ? "fallback" : "legacy_unknown"),
-    weeklyImpact: {
-      suggestedWeekAction,
-      suggestedNextCall: nextCall
-    },
-    createdAt,
-    updatedAt,
-    status: legacyStatus,
-    intentMatchStatus: legacyStatus,
-    executionScore: args.evidence.rulesSummary.executionScore,
-    executionScoreBand: args.evidence.rulesSummary.executionScoreBand,
-    executionScoreSummary: args.verdict?.sessionVerdict.summary ?? "Execution evidence is available for review.",
-    executionSummary: args.verdict?.explanation.whatHappened ?? "Execution evidence is available for review.",
-    summary: args.verdict?.sessionVerdict.summary ?? "Execution evidence is available for review.",
-    whyItMatters: args.verdict?.explanation.whyItMatters ?? "Use this review conservatively and in the context of the rest of the week.",
-    recommendedNextAction: args.verdict?.explanation.whatToDoNextTime ?? "Use one clear execution cue on the next similar session.",
-    diagnosisConfidence: args.evidence.rulesSummary.confidence,
-    executionScoreProvisional: args.evidence.rulesSummary.provisional,
-    suggestedWeekAdjustment: suggestedWeekAction,
-    evidence: args.verdict?.citedEvidence.flatMap((item) => item.support).slice(0, 4) ?? buildEvidenceSummary(args.evidence),
-    durationCompletion:
-      args.evidence.actual.durationSec && args.evidence.planned.durationSec
-        ? Number((args.evidence.actual.durationSec / args.evidence.planned.durationSec).toFixed(2))
-        : null,
-    intervalCompletionPct: args.evidence.actual.intervalCompletionPct,
-    timeAboveTargetPct: args.evidence.actual.timeAboveTargetPct,
-    avgHr: args.evidence.actual.avgHr,
-    avgPower: args.evidence.actual.avgPower,
-    avgIntervalPower: args.evidence.actual.avgIntervalPower ?? null,
-    normalizedPower: args.evidence.actual.normalizedPower ?? null,
-    trainingStressScore: args.evidence.actual.trainingStressScore ?? null,
-    intensityFactor: args.evidence.actual.intensityFactor ?? null,
-    totalWorkKj: args.evidence.actual.totalWorkKj ?? null,
-    avgCadence: args.evidence.actual.avgCadence ?? null,
-    avgPacePer100mSec: args.evidence.actual.avgPacePer100mSec ?? null,
-    bestPacePer100mSec: args.evidence.actual.bestPacePer100mSec ?? null,
-    avgStrokeRateSpm: args.evidence.actual.avgStrokeRateSpm ?? null,
-    maxStrokeRateSpm: args.evidence.actual.maxStrokeRateSpm ?? null,
-    avgSwolf: args.evidence.actual.avgSwolf ?? null,
-    elevationGainM: args.evidence.actual.elevationGainM ?? null,
-    elevationLossM: args.evidence.actual.elevationLossM ?? null,
-    poolLengthM: args.evidence.actual.poolLengthM ?? null,
-    lengthCount: args.evidence.actual.lengthCount ?? null,
-    hrZoneTimeSec: args.evidence.actual.hrZoneTimeSec ?? null,
-    paceZoneTimeSec: args.evidence.actual.paceZoneTimeSec ?? null,
-    maxHr: args.evidence.actual.maxHr ?? null,
-    maxPower: args.evidence.actual.maxPower ?? null,
-    firstHalfAvgHr: args.evidence.actual.splitMetrics?.firstHalfAvgHr ?? null,
-    lastHalfAvgHr: args.evidence.actual.splitMetrics?.lastHalfAvgHr ?? null,
-    firstHalfPaceSPerKm: args.evidence.actual.splitMetrics?.firstHalfPaceSPerKm ?? null,
-    lastHalfPaceSPerKm: args.evidence.actual.splitMetrics?.lastHalfPaceSPerKm ?? null,
-    firstHalfAvgCadence: args.evidence.actual.splitMetrics?.firstHalfAvgCadence ?? null,
-    lastHalfAvgCadence: args.evidence.actual.splitMetrics?.lastHalfAvgCadence ?? null,
-    firstHalfPacePer100mSec: args.evidence.actual.splitMetrics?.firstHalfPacePer100mSec ?? null,
-    lastHalfPacePer100mSec: args.evidence.actual.splitMetrics?.lastHalfPacePer100mSec ?? null,
-    firstHalfStrokeRate: args.evidence.actual.splitMetrics?.firstHalfStrokeRate ?? null,
-    lastHalfStrokeRate: args.evidence.actual.splitMetrics?.lastHalfStrokeRate ?? null,
-    executionCost: args.evidence.rulesSummary.executionCost,
-    missingEvidence: args.evidence.missingEvidence
-  };
-}
-
-export function parsePersistedExecutionReview(payload: Record<string, unknown> | null | undefined): PersistedExecutionReview | null {
-  if (!payload || typeof payload !== "object") return null;
-  if (payload.version === 2 && payload.deterministic && typeof payload.deterministic === "object") {
-    const narrativeSource: PersistedExecutionReview["narrativeSource"] =
-      payload.narrativeSource === "ai" || payload.narrativeSource === "fallback" || payload.narrativeSource === "legacy_unknown"
-        ? payload.narrativeSource
-        : "legacy_unknown";
-    return {
-      ...(payload as unknown as PersistedExecutionReview),
-      narrativeSource
-    };
-  }
-  return null;
-}
-
-export async function refreshObservedPatterns(supabase: SupabaseClient, athleteId: string) {
-  const { data: sessions, error } = await supabase
-    .from("sessions")
-    .select("id,date,execution_result")
-    .or(`athlete_id.eq.${athleteId},user_id.eq.${athleteId}`)
-    .not("execution_result", "is", null)
-    .order("date", { ascending: false })
-    .limit(30);
-
-  if (error) {
-    throw new Error(error.message);
-  }
-
-  const support = new Map<string, { label: string; detail: string; sourceSessionIds: string[] }>();
-  for (const session of sessions ?? []) {
-    const review = parsePersistedExecutionReview(session.execution_result as Record<string, unknown> | null);
-    if (!review) continue;
-    for (const issue of review.deterministic.detectedIssues) {
-      const existing = support.get(issue.code) ?? {
-        label: issue.code.replaceAll("_", " "),
-        detail: `This pattern has shown up across multiple reviewed sessions: ${issue.code.replaceAll("_", " ")}.`,
-        sourceSessionIds: []
-      };
-      existing.sourceSessionIds.push(session.id);
-      support.set(issue.code, existing);
-    }
-  }
-
-  const repeated = [...support.entries()].filter(([, value]) => value.sourceSessionIds.length >= 2);
-  for (const [patternKey, value] of repeated) {
-    const supportCount = value.sourceSessionIds.length;
-    const confidence = supportCount >= 4 ? "high" : supportCount >= 3 ? "medium" : "low";
-    const { error: upsertError } = await supabase.from("athlete_observed_patterns").upsert({
-      athlete_id: athleteId,
-      pattern_key: patternKey,
-      label: value.label,
-      detail: value.detail,
-      support_count: supportCount,
-      confidence,
-      last_observed_at: new Date().toISOString(),
-      source_session_ids: value.sourceSessionIds
-    }, {
-      onConflict: "athlete_id,pattern_key"
-    });
-    if (upsertError) {
-      throw new Error(upsertError.message);
-    }
-  }
-}
-
-export async function buildWeeklyExecutionBrief(args: {
-  supabase: SupabaseClient;
-  athleteId: string;
-  weekStart: string;
-  weekEnd: string;
-  athleteContext: AthleteContextSnapshot | null;
-  extraActivityCount?: number;
-}) {
-  const { data: sessions, error } = await args.supabase
-    .from("sessions")
-    .select("id,session_name,type,date,execution_result")
-    .or(`athlete_id.eq.${args.athleteId},user_id.eq.${args.athleteId}`)
-    .gte("date", args.weekStart)
-    .lte("date", args.weekEnd)
-    .not("execution_result", "is", null)
-    .order("date", { ascending: true });
-
-  if (error) {
-    throw new Error(error.message);
-  }
-
-  const reviews = (sessions ?? [])
-    .map((session) => {
-      const review = parsePersistedExecutionReview(session.execution_result as Record<string, unknown> | null);
-      if (!review) return null;
-      return {
-        id: session.id,
-        name: (session.session_name ?? session.type ?? "Session").trim(),
-        review
-      };
-    })
-    .filter((item): item is { id: string; name: string; review: PersistedExecutionReview } => Boolean(item));
-
-  const reviewedCount = reviews.length;
-  const onTargetCount = reviews.filter((item) => item.review.deterministic.rulesSummary.intentMatch === "on_target").length;
-  const partialCount = reviews.filter((item) => item.review.deterministic.rulesSummary.intentMatch === "partial").length;
-  const missedCount = reviews.filter((item) => item.review.deterministic.rulesSummary.intentMatch === "missed").length;
-  const provisionalCount = reviews.filter((item) => item.review.deterministic.rulesSummary.provisional).length;
-
-  const sessionsNeedingAttention = reviews
-    .filter((item) => item.review.deterministic.rulesSummary.intentMatch !== "on_target")
-    .slice(0, 3)
-    .map((item) => ({
-      sessionId: item.id,
-      sessionName: item.name,
-      scoreHeadline: item.review.verdict?.sessionVerdict.headline ?? item.review.executionScoreSummary,
-      reason: item.review.verdict?.explanation.whyItMatters ?? item.review.whyItMatters
-    }));
-
-  const keyPositive = onTargetCount > 0 ? `${onTargetCount} reviewed session${onTargetCount === 1 ? "" : "s"} are landing on target.` : null;
-  const keyRisk = sessionsNeedingAttention[0]?.reason ?? null;
-  const nextWeekDecision =
-    missedCount > 0
-      ? "Keep the next key session controlled and protect recovery rather than adding load."
-      : partialCount > 0
-        ? "Progress only if the next key session lands cleanly."
-        : "Keep the current structure and carry the same execution control into next week.";
-
-  const contextCue = args.athleteContext?.declared.weeklyConstraints[0] ?? args.athleteContext?.declared.limiters[0]?.value ?? null;
-  const extraCount = args.extraActivityCount ?? 0;
-  const extraNote = extraCount > 0 ? ` ${extraCount} extra session${extraCount === 1 ? "" : "s"} also logged this week.` : "";
-
-  return {
-    weekHeadline:
-      reviewedCount === 0
-        ? extraCount > 0
-          ? `${extraCount} extra session${extraCount === 1 ? "" : "s"} logged — reviews building`
-          : "Reviews are still building, so keep the week steady"
-        : missedCount > 0
-          ? "Execution is mostly on track, but one key session came up short"
-          : partialCount > 0
-            ? "Execution is on track overall, with a few sessions needing attention"
-            : "Execution is on track this week",
-    weekSummary:
-      reviewedCount === 0
-        ? `Early uploads are in.${extraNote} Hold the current structure for now, then let the next reviewed sessions sharpen the call.`
-        : `${extraCount > 0 ? `${extraCount} extra ${extraCount === 1 ? "session" : "sessions"} also logged this week.` : "Execution reviewed against this week's training intent."}${contextCue ? ` ${contextCue} noted.` : ""}`,
-    keyPositive,
-    keyRisk,
-    nextWeekDecision,
-    trend: {
-      reviewedCount,
-      onTargetCount,
-      partialCount,
-      missedCount,
-      provisionalCount
-    },
-    sessionsNeedingAttention,
-    confidenceNote:
-      provisionalCount > 0
-        ? `${provisionalCount} review${provisionalCount === 1 ? "" : "s"} are still provisional because evidence is incomplete.`
-        : null
-  } satisfies WeeklyExecutionBrief;
-}
